refactor(demo): move app guards into DemoModule (ADR-0001)
AppAdminGuard/AppUserGuard are demo app authz, not IdP primitives; they
move next to the demo RolesService they depend on. Fixed relative import
paths: guards now import RolesService from ../roles/ instead of
../demo/roles/, and importers updated to find guards at demo/guards/.
