+++ api/src/demo/guards/app-admin.guard.ts
@@ -1,18 +1,22 @@
 import { Injectable, CanActivate, ExecutionContext, ForbiddenException, Logger } from '@nestjs/common';
-import { RolesService } from '../demo/roles/roles.service';
+import { Reflector } from '@nestjs/core';
+import { RolesService } from '../roles/roles.service';
 
 @Injectable()
-export class AppUserGuard implements CanActivate {
-  private readonly logger = new Logger(AppUserGuard.name);
+export class AppAdminGuard implements CanActivate {
+  private readonly logger = new Logger(AppAdminGuard.name);
 
-  constructor(private rolesService: RolesService) { }
+  constructor(
+    private reflector: Reflector,
+    private rolesService: RolesService,
+  ) { }
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
     const request = context.switchToHttp().getRequest();
     const user = request.user;
 
     if (!user || !user.userId) {
-      this.logger.warn('AppUserGuard: No user in request');
+      this.logger.warn('AppAdminGuard: No user in request');
       throw new ForbiddenException('User not authenticated');
     }
 
@@ -22,10 +26,9 @@ export class AppUserGuard implements CanActivate {
       appRole = await this.rolesService.getAppRole(user.userId);
     }
 
-    // Both app_admin and app_user can access endpoints protected by AppUserGuard
-    if (appRole !== 'app_admin' && appRole !== 'app_user') {
-      this.logger.warn(`AppUserGuard: User ${user.userId} has invalid app role: ${appRole}`);
-      throw new ForbiddenException(`Access denied. Invalid app role: ${appRole}`);
+    if (appRole !== 'app_admin') {
+      this.logger.warn(`AppAdminGuard: User ${user.userId} has role ${appRole}, required app_admin`);
+      throw new ForbiddenException(`Access denied. Required app role: app_admin. Your app role: ${appRole}`);
     }
 
     // Add appRole to user object for use in controllers
