+++ api/src/demo/guards/app-user.guard.ts
@@ -1,22 +1,18 @@
 import { Injectable, CanActivate, ExecutionContext, ForbiddenException, Logger } from '@nestjs/common';
-import { Reflector } from '@nestjs/core';
-import { RolesService } from '../demo/roles/roles.service';
+import { RolesService } from '../roles/roles.service';
 
 @Injectable()
-export class AppAdminGuard implements CanActivate {
-  private readonly logger = new Logger(AppAdminGuard.name);
+export class AppUserGuard implements CanActivate {
+  private readonly logger = new Logger(AppUserGuard.name);
 
-  constructor(
-    private reflector: Reflector,
-    private rolesService: RolesService,
-  ) { }
+  constructor(private rolesService: RolesService) { }
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
     const request = context.switchToHttp().getRequest();
     const user = request.user;
 
     if (!user || !user.userId) {
-      this.logger.warn('AppAdminGuard: No user in request');
+      this.logger.warn('AppUserGuard: No user in request');
       throw new ForbiddenException('User not authenticated');
     }
 
@@ -26,9 +22,10 @@ export class AppAdminGuard implements CanActivate {
       appRole = await this.rolesService.getAppRole(user.userId);
     }
 
-    if (appRole !== 'app_admin') {
-      this.logger.warn(`AppAdminGuard: User ${user.userId} has role ${appRole}, required app_admin`);
-      throw new ForbiddenException(`Access denied. Required app role: app_admin. Your app role: ${appRole}`);
+    // Both app_admin and app_user can access endpoints protected by AppUserGuard
+    if (appRole !== 'app_admin' && appRole !== 'app_user') {
+      this.logger.warn(`AppUserGuard: User ${user.userId} has invalid app role: ${appRole}`);
+      throw new ForbiddenException(`Access denied. Invalid app role: ${appRole}`);
     }
 
     // Add appRole to user object for use in controllers
