+++ api/src/demo/roles/roles.controller.ts
@@ -11,7 +11,7 @@ import {
 } from '@nestjs/common';
 import { RolesService } from './roles.service';
 import { AuthenticatedGuard } from '../../guards/authenticated.guard';
-import { AppAdminGuard } from '../../guards/app-admin.guard';
+import { AppAdminGuard } from '../guards/app-admin.guard';
 import { RoleGuard } from '../../guards/role.guard';
 import { RequireRole } from '../../decorators/require-role.decorator';
 import { Public } from '../../decorators/public.decorator';
